refactor: move formatDiffWithLineNumbers to a new module
